feat: jwt authentication & role authentication

+++ Controllers/ScheduleController.cs
@@ -1,7 +1,11 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using ShiftPro.Data;
 using ShiftPro.Dtos;
 using ShiftPro.Interfaces;
+using System.Security.Claims;
 
 namespace ShiftPro.Controllers
 {
@@ -10,11 +14,14 @@ namespace ShiftPro.Controllers
     public class ScheduleController : ControllerBase
     {
         private readonly IScheduleService _service;
-        public ScheduleController(IScheduleService service)
+        private readonly AppDbContext _context;
+        public ScheduleController(IScheduleService service, AppDbContext context)
         {
             _service = service;
+            _context = context;
         }
 
+        [Authorize(Roles = "Admin,Boss")]
         [HttpGet]
         public async Task<IActionResult> GetAllSchedules()
         {
@@ -27,21 +34,33 @@ namespace ShiftPro.Controllers
             return Ok(result);  
         }
 
-        [HttpGet("{id}")]
-        public async Task<IActionResult> GetAllSchedules([FromRoute] int id)
+        [Authorize]
+        [HttpGet("me")]
+        public async Task<IActionResult> GetAllScheduleById()
         {
-            var result = await _service.GetSchedulesById(id);
+            var user = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            var userId = int.Parse(user);
+
+            var result = await _service.GetSchedulesById(userId);
 
             if (result == null)
             {
-                return NotFound("找不到此Schedule");
+                return NotFound("尚無排班紀錄");
             }
             return Ok(result);
         }
 
+        [Authorize(Roles = "Employee")]
         [HttpPost]
         public async Task<IActionResult> CreateSchedule([FromBody] CreateScheduleDto dto)
         {
+            var user = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            var userId = int.Parse(user);
+
+            if (dto.EmployeeId != userId)
+            {
+                return Forbid();
+            }
             var result =await _service.CreateSchedule(dto);
 
             if (result == null)
@@ -54,9 +73,26 @@ namespace ShiftPro.Controllers
             });
         }
 
+        [Authorize(Roles = "Employee,Admin")]
         [HttpPatch("{id}")]
         public async Task<IActionResult> UpdateSchedule([FromRoute] int id, [FromBody] UpdateScheduleDto dto)
         {
+            var user = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            var userId = int.Parse(user);
+
+
+           var schedule =await  _context.Schedules.FirstOrDefaultAsync(x => x.Id == id);
+
+            if (schedule == null)
+            {
+                return NotFound("查無Schedule資料");
+            }
+
+            if (schedule.EmployeeId != userId && !User.IsInRole("Admin"))
+            {
+                return Forbid();
+            }
+
             var result = await _service.UpdateSchedule(id, dto);
 
             if (result == false)
@@ -69,9 +105,25 @@ namespace ShiftPro.Controllers
             });
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteSchedule([FromRoute] int id)
         {
+            var user = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            var userId = int.Parse(user);
+
+            var schedule = await _context.Schedules.FirstOrDefaultAsync(x => x.Id == id);
+
+            if (schedule == null)
+            {
+                return NotFound("查無Schedule資料");
+            }
+
+            if (schedule.EmployeeId != userId && !User.IsInRole("Admin"))
+            {
+                return Forbid();
+            }
+
             var result =await _service.DeleteSchedule(id);
 
             if (result == false)
@@ -84,6 +136,7 @@ namespace ShiftPro.Controllers
             });
         }
 
+        [Authorize(Roles = "Boss")]
         [HttpGet("Monthly")]
         public async Task<IActionResult> GetMonthlySchedules([FromQuery] int year, [FromQuery] int month)
         {
@@ -97,7 +150,7 @@ namespace ShiftPro.Controllers
                 return Ok( result );
             }
 
-
+        [Authorize(Roles = "Boss")]
         [HttpGet("EmployeeReport")]
         public async Task<IActionResult> GetEmployeesMonthlyCount([FromQuery] int year, [FromQuery] int month)
         {
